Move facebook-loader.js into its component directory
