+++ extensions/amp-facebook/0.1/facebook-loader.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {htmlFor} from '../../src/static-template';
+import {htmlFor} from '../../../src/static-template';
 
 /**
  * Common function to create the facebook loader logo for all amp-facebook-*
